Deque Finished
I have added the trim to deque. I have fixed the bugs from free, clear,
findn_cyclical

Issues to close when excepted:
HKL-9
HKL-8

+++ src/hkl_deque.c
@@ -218,7 +218,7 @@ void hkl_deque_trim(HklDeque *deque)
   {
     void ** tmp = malloc(deque->size * sizeof(void*));
 
-    for(int i = 0; i < deque->size; ++i)
+    for(int i = 0; i < (int)deque->size; ++i)
     {
       tmp[i] = hkl_deque_findn(deque, i);
     }
